MAINT: refactoring
