+++ enigma-principal/app/src/boot_network/mod.rs
@@ -1,3 +1,2 @@
 pub mod principal_manager;
-pub mod enigma_contract;
 pub mod principal_utils;
