+++ enigma-tools-u/src/lib.rs
@@ -4,14 +4,21 @@
 extern crate failure;
 extern crate reqwest;
 extern crate serde_json;
-
 #[macro_use]
 extern crate serde_derive;
 extern crate serde;
 extern crate rlp;
+// webv3 utils 
+extern crate web3;
+extern crate libc;
+extern crate rustc_hex;
+extern crate tiny_keccak;
+// enigma contract 
+extern crate sgx_types;
+
 pub mod attestation_service;
 pub mod common_u;
-
+pub mod web3_utils;
 
 #[cfg(test)]
 mod tests {
